+++ milvus/MilvusClient.ts
@@ -1,7 +1,31 @@
-import { GRPCClient, ClientConfig } from '.';
 import { ChannelOptions } from '@grpc/grpc-js';
+import { GRPCClient, ClientConfig } from '.';
+import sdkInfo from '../sdk.json';
 
+/**
+ * Milvus Client class that extends GRPCClient and handles communication with Milvus server.
+ */
 export class MilvusClient extends GRPCClient {
+  /**
+   * Returns the SDK information.
+   * SDK information will be generated on the building phase
+   * @returns Object containing SDK version and recommended Milvus version.
+   */
+  static get sdkInfo() {
+    return {
+      version: sdkInfo.version,
+      recommandMilvus: sdkInfo.milvusVersion,
+    };
+  }
+
+  /**
+   * Creates a new instance of MilvusClient.
+   * @param configOrAddress The Milvus server's address or client configuration object.
+   * @param ssl Whether to use SSL or not.
+   * @param username The username for authentication.
+   * @param password The password for authentication.
+   * @param channelOptions Additional channel options for gRPC.
+   */
   constructor(
     configOrAddress: ClientConfig | string,
     ssl?: boolean,
@@ -9,13 +33,15 @@ export class MilvusClient extends GRPCClient {
     password?: string,
     channelOptions?: ChannelOptions
   ) {
+    // setup the configuration
     super(configOrAddress, ssl, username, password, channelOptions);
-
-    // connect
+    // connect here
     this.connect();
   }
 
-  // overload
+  /**
+   * Overloads the original connect function in GRPCClient to add additional functionality.
+   */
   connect() {
     super.connect();
   }
