+++ milvus/grpc/Collection.ts
@@ -1,3 +1,4 @@
+import path from 'path';
 import {
   promisify,
   formatKeyValueData,
@@ -5,7 +6,7 @@ import {
   checkCollectionFields,
   checkCollectionName,
   sleep,
-} from '../utils';
+} from '../../utils';
 import { BaseClient } from './BaseClient';
 import {
   ERROR_REASONS,
@@ -42,8 +43,10 @@ import {
   ReplicasResponse,
   GetLoadingProgressResponse,
   GetLoadStateResponse,
-} from '.';
-import { assignTypeParams } from '../utils';
+} from '..';
+import { assignTypeParams } from '../../utils';
+
+const schemaProtoPath = path.resolve(__dirname, '../proto/proto/schema.proto');
 
 /**
  * @see [collection operation examples](https://github.com/milvus-io/milvus-sdk-node/blob/main/example/Collection.ts)
