+++ milvus/grpc/User.ts
@@ -1,4 +1,4 @@
-import { promisify, stringToBase64 } from '../utils';
+import { promisify, stringToBase64 } from '../../utils';
 import { Resource } from './Resource';
 import {
   ERROR_REASONS,
@@ -25,7 +25,7 @@ import {
   SelectUserResponse,
   SelectGrantResponse,
   HasRoleResponse,
-} from '.';
+} from '..';
 
 export class User extends Resource {
   /**
