+++ milvus/index.ts
@@ -1,17 +1,9 @@
 // consts
-export * from './const/ErrorReason';
-export * from './const/Milvus';
+export * from './const';
+// utils
+export * from '../utils';
 // types
-export * from './types/Collection';
-export * from './types/Data';
-export * from './types/Common';
-export * from './types/Index';
-export * from './types/Partition';
-export * from './types/Response';
-export * from './types/User';
-export * from './types/Resource';
-export * from './types/Client';
-// client
-export * from './BaseClient';
-export * from './GrpcClient';
+export * from './types';
+// clients
+export * from './grpc/GrpcClient';
 export * from './MilvusClient';
