+++ milvus/types.ts
@@ -1,17 +1,10 @@
+// @deprecated
 export * from './types/Collection';
-
 export * from './types/Data';
-
 export * from './types/Common';
-
-export * from './types/Index';
-
+export * from './types/MilvusIndex';
 export * from './types/Partition';
-
 export * from './types/Response';
-
 export * from './types/User';
-
 export * from './types/Resource';
-
 export * from './types/Client';
